Пофиксил часть багов, которые resolved в pull requests.
