+++ src/main/java/ru/sbt/mipt/oop/commands/CommandSender.java
@@ -1,9 +1,6 @@
 package ru.sbt.mipt.oop.commands;
 
-import ru.sbt.mipt.oop.commands.CommandFactory;
-import ru.sbt.mipt.oop.commands.SensorCommand;
-
-public class SensorCommandFactory implements CommandFactory {
+public class CommandSender implements CommandFactory {
     @Override
     public void sendCommand(SensorCommand command) {
         System.out.println("Pretend we're sending command " + command);
